Move serve tests
