+++ tests/core/serve/data/cat.jpg
