+++ tests/core/serve/data/fish.jpg
