+++ tests/core/serve/data/number5.png
