+++ tests/core/serve/test_compat/test_cached_property.py
