+++ tests/core/serve/test_types/test_bbox.py
