+++ tests/core/serve/test_types/test_image.py
