+++ tests/core/serve/test_types/test_text.py
